Passing All Tests - Refactored the test suite in userprofile app to have a base Test Case - Wrote tests we were lacking on the profile model
